Update dev dependencies for and builds of custom JS code
In JS dev dependency pbk-custom.js, update the markup that is inserted to create the page header on news pages. Next, update the submodule holding universal dev dependencies used to build custom JS code in every DAESA website development project. Finally, rebuild the files containing custom JS code that is ready to be deployed on the website associated with this project.